Update post infrastructure to use long

+++ src/back/Plants.Application/Commands/DeletePostCommand.cs
@@ -2,5 +2,5 @@
 
 namespace Plants.Application.Commands;
 
-public record DeletePostCommand(int PostId) : IRequest<DeletePostResult>;
+public record DeletePostCommand(long PostId) : IRequest<DeletePostResult>;
 public record DeletePostResult(bool Deleted);
