+++ src/back/Plants.Application/Contracts/IPostService.cs
@@ -5,7 +5,7 @@ namespace Plants.Application.Contracts;
 
 public interface IPostService
 {
-    Task<PostResultItem?> GetBy(int postId);
-    Task<PlaceOrderResult> Order(int postId, string city, int mailNumber);
-    Task<DeletePostResult> Delete(int postId);
+    Task<PostResultItem?> GetBy(long postId);
+    Task<PlaceOrderResult> Order(long postId, string city, int mailNumber);
+    Task<DeletePostResult> Delete(long postId);
 }
